+++ ReleaseNotes/release_notes.md
@@ -2,6 +2,7 @@
 
   1. add new module `ostap.io.zstshelve` with shelve-like databse using `zstandard` compression 
   1. add generalized Pareto distribution and reparameterised version of the exponentiated generalized Pareto distribution: functions and PDFs
+  1. add generalized extreme value distribution: function and PDF
  
 ## Backward incompatible: